Split out Repository & repository_url_validator
